requirements.txt
PyYAML==5.1.2
Flask==1.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.5.2
PyMySQL==0.9.2